Task 1: extract silentFetch into shared Helpers
Move silentFetch out of Macros.tsx into components/Helpers/http.ts so
both the existing manual URI_SILENT/[SILENT] macro paths and the
upcoming event-macros engine (Task 2+) can share it. Adds an optional
SilentFetchOptions (timeoutMs, onSettled) for the in-flight-tracking
needs of Task 2 - behavior-preserving for existing callers, which pass
no options and get the byte-identical pre-extraction fetch() call.
